Add a notebook for launching from the moon
This notebook complements the previous notebook for landing on the moon
very nicely.

What's also nice about it is that it offers a comparison between direct
and indirect methods of optimization. It is a direct approach to an
example in Optimal Control with Aerospace Applications by Longuski et
al. which is solved by indirect means in the book.